requirements.txt
discord.py==2.3.1
SQLAlchemy==2.0.18
mysql-connector-python==8.1.0 
